Add node_ids function to NodePopulation
From the specs, node ids (or label) can be different from the positional
indices. See:
https://github.com/AllenInstitute/sonata/blob/master/docs/SONATA_DEVELOPER_GUIDE.md#nodes---optional-reserved-attributes

The mapping between the positional indices and label indices is done via
the "node_id" field if this one is present in the population. If the
field is not present then: positional indices == label indices which is
the case in our circuits today but may change one day.

With complete circuits, this is important to have access to this field
since source/target_node_id returns label indices and not the positional
indices.