Return wrapper numpy arrays for member variables
We finally solve the problem of array data stored somewhere and
we want to return only a wrapper numpy array. Numpy accepts a handle
object for Reference count adjustment.
It's nevertheless weird there is no proper documentation or pybind11
API to retrieve the handle/python object of the current C++ object.