API for configuring HDF5 properties.
This commit introduces a class `IoOpts` which allows users to inject HDF5
properties into the File Access Property List and the Data Transfer Property
List.

These are needed to get access to performance critical options in HDF5. For
example, paged HDF5 files, or MPI-IO.

These can be extended to allow configuring the reading algorithms. For example
the block size when aggregating and whether aggregation is permitted to
generate a variable number of HDF5 API calls.