Collective MPI-IO.
This commit implements an `IoOpts` for performing (collective) MPI-IO. In this
form it (unconditionally) adds the `mpi4py` as a build requirement to
`libsonata`. Due to packaging issues, it's not possible to have an optional
build-dependency. Therefore, this commit would be moved to a new package
`libsonata-mpi` which can be made an optional dependency of `libsonata`.